src/main/java/com/example/socialBookStore/controllers/NotificationController.java
package com.example.socialBookStore.controllers;


import com.example.socialBookStore.services.NotificationService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/notifications")
public class NotificationController {

    @Autowired
    private NotificationService notificationService;

    @PostMapping("/delete/{id}")
    public ResponseEntity<?> deleteNotification(@PathVariable Integer id){
        try {
            notificationService.deleteNotificationById(id);
            return ResponseEntity.ok("Notification deleted successfully");
        } catch (Exception e) {
            return ResponseEntity.badRequest().body("An error occured with notification deletion");
        }


    }
}
